convert to maven project [WIP]
