+++ .gitignore
@@ -7,3 +7,4 @@
 */*/.DS_Store
 */*/*/.DS_Store
 */*/*/*/.DS_Store
+/target/
