--- .project
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<projectDescription>
-	<name>ScoreboardUUID</name>
-	<comment>Minecraft Plugin</comment>
-	<projects>
-	</projects>
-	<buildSpec>
-		<buildCommand>
-			<name>org.eclipse.jdt.core.javabuilder</name>
-			<arguments>
-			</arguments>
-		</buildCommand>
-	</buildSpec>
-	<natures>
-		<nature>org.eclipse.jdt.core.javanature</nature>
-	</natures>
-	<linkedResources>
-		<link>
-			<name>EvLib</name>
-			<type>2</type>
-			<location>/Users/Nate/Dropbox/Minecraft/Development/Plugins/EvLib/src</location>
-		</link>
-	</linkedResources>
-</projectDescription>
