--- config.yml
@@ -1,6 +1,0 @@
-# Scores to updatd when a player changes their name
-uuid-based-scores:
-  - deaths
-  - 'num-advancements'
-  - buildscore
-  - votes
