--- plugin.yml
@@ -1,18 +1,0 @@
-name: ScoreboardUUID
-main: net.evmodder.ScoreboardUUID.ScoreboardUUID
-author: EvModder/EvDoc
-version: 1.0 #1.0=initial release
-api-version: 1.13
-
-commands:
-   about:
-      description: View information about a player
-      usage: /about <name>
-      aliases: [seen,info]
-      permission: hardcore.about
-      permission-message: You do not have permission to do this.
-
-permissions:
-   hardcore.about:
-      description: Permission to use /about
-      default: true
